# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "The APIs provided by the NDR component",
    "termsOfService" : "https://github.com/Montimage/maip",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.com"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    }
  ],
...
  "paths" : {
...
    "/predict/" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "The prediction process status",
        "description" : "Get the status of the latest prediction process",
        "operationId" : "getPredictionStatus",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/predictingStatus"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "predictions" ],
        "summary" : "Start predicting",
        "description" : "Request predicting for an input traffic",
        "operationId" : "requestPredicting",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/predictConfig"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/predictingStatus"
                }
              }
            }
          },
          "500" : {
            "description" : "Failed to process a prediction"
          }
        }
      }
    },
...
    "/predict/offline" : {
      "post" : {
        "tags" : [ "predictions" ],
        "summary" : "Queue-based offline prediction",
        "description" : "Start offline prediction using job queue for non-blocking execution",
        "operationId" : "predictOffline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "modelId" : {
                    "type" : "string",
                    "example" : "6d9aebed-aa34-47c5-be80-b41280c32163.h5"
                  },
                  "reportId" : {
                    "type" : "string",
                    "example" : "report-honeypot.pcap-342727c9"
                  },
                  "reportFileName" : {
                    "type" : "string",
                    "example" : "1675893004.993760_0_honeypot.pcap.csv"
                  },
                  "useQueue" : {
                    "type" : "boolean",
                    "example" : true
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "jobId" : {
                      "type" : "string"
                    },
                    "predictionId" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/predict/job/{jobId}" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Get queued prediction job status",
        "description" : "Get the status of a queued prediction job",
        "operationId" : "getPredictionJobStatus",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "description" : "Job ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/jobStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/security/rule-based/online/start" : {
      "post" : {
        "tags" : [ "security" ],
        "summary" : "Start online rule-based detection",
        "description" : "Start real-time rule-based intrusion detection",
        "operationId" : "startRuleBasedOnline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "iface" : {"type" : "string"},
                  "intervalSec" : {"type" : "number"},
                  "verbose" : {"type" : "boolean"},
                  "excludeRules" : {"type" : "array", "items" : {"type" : "string"}},
                  "cores" : {"type" : "number"}
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Successfully started"
          }
        }
      }
    },
...
    "/online/start" : {
      "post" : {
        "tags" : [ "online" ],
        "summary" : "Start online packet capture",
        "description" : "Start capturing packets with tcpdump",
        "operationId" : "startOnlineCapture",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "iface" : {"type" : "string"},
                  "windowSec" : {"type" : "number"},
                  "totalDurationSec" : {"type" : "number"},
                  "filter" : {"type" : "string"}
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Capture started"
          }
        }
      }
    },
...
    "/queue/status/{queueName}/{jobId}" : {
      "get" : {
        "tags" : [ "queue" ],
        "summary" : "Get job status",
        "description" : "Get status of a queued job",
        "operationId" : "getJobStatus",
        "parameters" : [ {
          "name" : "queueName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Job status retrieved"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "predictingStatus" : {
        "type" : "object",
        "properties" : {
          "isRunning" : {
            "type" : "boolean",
            "example" : true
          },
          "lastPredictedId" : {
            "type" : "string",
            "example" : "5f4ffd91-69ba-4f1c-985d-c3a7f238e8b5"
          },
          "config" : {
            "$ref" : "#/components/schemas/predictConfig"
          },
          "lastPredictedAt" : {
            "type" : "string",
            "example" : "1694705577544"
          }
        }
      },
...
      "predictConfig" : {
        "type" : "object",
        "properties" : {
          "predictConfig" : {
            "type" : "object",
            "properties" : {
              "modelId" : {
                "type" : "string",
                "example" : "6d9aebed-aa34-47c5-be80-b41280c32163.h5"
              },
              "inputTraffic" : {
                "type" : "object",
                "properties" : {
                  "type" : {
                    "type" : "string",
                    "example" : "report"
                  },
                  "value" : {
                    "type" : "object",
                    "properties" : {
                      "reportId" : {
                        "type" : "string",
                        "example" : "report-honeypot.pcap-342727c9-e253-4124-85fd-974c2ad60bae"
                      },
                      "reportFileName" : {
                        "type" : "string",
                        "example" : "1675893004.993760_0_honeypot.pcap.csv"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "jobStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["completed", "failed", "processing", "waiting"],
            "example": "completed"
          },
          "result": {
            "type": "object",
            "description": "Job result data when completed"
          },
          "failedReason": {
            "type": "string",
            "description": "Error message if job failed"
          },
          "progress": {
            "type": "number",
            "description": "Job progress percentage (0-100)"
          }
        }
      }
    }
  }
}